requirements.txt
fastapi
uvicorn
beanie
motor
pydantic[email]
google-api-python-client
google-auth-oauthlib
pyjwt
python-multipart
reportlab
dnspython
python-dotenv
